// src/app/shared/enum/reports-details.enum.ts
export enum ReportsDetails {
  MENU_SALES_TYPE = 'tipos-de-venda-cardapio',
  MENU_PAYMENT_METHOD = 'tipos-de-pagamento-cardapio',
  CONVERSION = 'detalhes-conversao',
  SALES_BY_USER = 'vendas-por-usuario',
  DELIVERY_BY_DELIVERYMAN = 'entregas-por-entregadores',
  ALL_PRODUCTS_SALE = 'todos-os-produtos',
  SALES_BOX_SALES_TYPE = 'tipos-de-venda',
  SALES_BOX_PAYMENT_METHOD = 'tipos-de-pagamento',
}